Separate node and core in project
