+++ jest/setup.ts
@@ -1,4 +1,4 @@
-import { setupEnvironment } from "../src/env";
+import { setupEnvironment } from "../src/core/env";
 import { ReadableStream } from "web-streams-polyfill/ponyfill/es2018";
 import { TextDecoder } from "util";
