+++ src/examples/simple-node-server.ts
@@ -1,4 +1,4 @@
-import { HttpRequest, HttpResponse, HttpStatus } from "../http";
+import { HttpRequest, HttpResponse, HttpStatus } from "../core/http";
 import * as http from "http";
 
 import { toNodeRequestListener } from "../node/server";
