+++ src/node/__tests__/HttpBodyImpl.spec.ts
@@ -1,5 +1,5 @@
 import { HttpBodyImpl } from "../HttpBodyImpl";
-import { stringToReadableStream } from "../../utils/stringToReadableStream";
+import { stringToReadableStream } from "../../core/utils";
 
 describe("HttpBodyImpl", () => {
   it("should be return json", async () => {
