+++ src/node/__tests__/server.spec.ts
@@ -1,8 +1,8 @@
 import * as http from "http";
 import { get } from "request-promise";
-import { HttpResponse } from "../../http";
+import { HttpResponse } from "../../core/http";
 import { toNodeRequestListener } from "../server";
-import { HttpHandler } from "../../http4ts";
+import { HttpHandler } from "../../core/http4ts";
 import { HttpBodyImpl } from "../HttpBodyImpl";
 
 async function runOnTestServer(
